119 - passes in object_key instead of original_filename
Before we were passing in the original filename, and letting
the remediation job logic generate an object key. However, since the object
key contains randomized characters, it was ending up different than
the object key generated in the jobs_controller to create the presigned_url
for uploading. Passing in an object key to the remediation job elimates
  this discrepancy.